[ser] address Broadcom ser test issues (#2040)
- When using command() instead of shell(), we got host unreachable error.
- With shell() we got pass in seconds, which means that the test didn't run.
- With specifying executable, the test actually runs.
- Add cleanup code since this test has capability of leaving DUT in bad state.
- Move ssh timeout handling to ser test.
- Reboot to forcefully reset connection.
